% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq_create_label.R
\name{eq_create_label}
\alias{eq_create_label}
\title{Leaflet Map Annotation Label}
\usage{
eq_create_label(data)
}
\arguments{
\item{data}{An NOAA earthquake data file.}
}
\description{
\code{eq_create_label} generates a more informative annotation label for the leaflet map generated
with \code{\link{eq_map}}.
}
\details{
The fields displayed are LOCATION, SIZE, and DEATHS
}
\examples{

\dontrun{
data(earthquakes)
earthquakes \%>\%
eq_clean_data() \%>\%
dplyr::filter(COUNTRY == "MEXICO" & lubridate::year(DATE) >= 2000) \%>\%
dplyr::mutate(popup_text = eq_create_label(.)) \%>\%
eq_map(annot_col = "popup_text")
}

}
