% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq_location_clean.R
\name{eq_location_clean}
\alias{eq_location_clean}
\title{Clean LOCATION Variable}
\usage{
eq_location_clean(rawdata)
}
\arguments{
\item{rawdata}{An NOAA tab-delimited earthquake file.}
}
\value{
A dataframe.
  The dataframe returned is processed to be used with \code{\link{geom_timeline}} and \code{\link{geom_timeline_label}}.
}
\description{
\code{eq_location_clean} cleans the \code{LOCATION_NAME} variable in an NOAA earthquakes data file.
Country name and white-space is removed, and the name is converted to title-case.
}
\examples{

\dontrun{
data(earthquakes)
earthquakes_cleaned <- earthquakes \%>\% eq_clean_data() \%>\% eq_location_clean()
}

}
