% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq_map.R
\name{eq_map}
\alias{eq_map}
\title{Earthquake Leaflet Map}
\usage{
eq_map(data, annot_col = "DATE")
}
\arguments{
\item{data}{An NOAA earthquakes data file}

\item{annot_col}{Character. The name of the variable to display in the annotation.}
}
\value{
An interactive leaflet map displaying earthquake locations.
}
\description{
\code{eq_map} generates a \code{\link[leaflet]{leaflet}} map displaying the geographical position of earthquakes
in a NOAA earthquakes data file.
}
\details{
The size of the circles displayed are proportional to the earthquake magnitude.
Clicking on a given circle will display the associated value of the variable specified by \code{annot_col}.
}
\examples{

\dontrun{
data(earthquakes)
earthquakes \%>\%
eq_clean_data() \%>\%
dplyr::filter(COUNTRY == "MEXICO" & lubridate::year(DATE) >= 2000) \%>\%
eq_map(annot_col = "DATE")
}

}
