% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_timeline.R
\name{geom_timeline}
\alias{geom_timeline}
\title{Earthquake Timeline}
\usage{
geom_timeline(mapping = NULL, data = NULL, stat = "identity",
  position = "identity", na.rm = FALSE, show.legend = NA,
  inherit.aes = TRUE, ...)
}
\arguments{
\item{mapping}{A set of aesthetics created by \code{aes()}.}

\item{data}{The NOAA earthquake data to be displayed.}

\item{stat}{The statistical transformation for that layer.}

\item{position}{Position adjustment.}

\item{na.rm}{Option to remove missing values.}

\item{show.legend}{Option to display legends.}

\item{inherit.aes}{Option to override the default aesthetics.}

\item{...}{Other arguments passed to \code{layer()}.}
}
\description{
\code{geom_timeline} displays a timeline of earthquakes. Date is displayed on the x-axis and countries are displayed
on the y-axis as individual timelines. The size of the point is proportional to earthquake magnitude, and the color
is related to the number of deaths resulting from the earthquake.
}
\section{Aesthetics}{

\code{geom_timeline} can be used with the following aesthetics:
\itemize{
 \item x
 \item y
 \item size
 \item color
 \item alpha
 \item shape
 \item stroke
 \item fill
}
}

\examples{

\dontrun{
data(earthquakes)
earthquakes \%>\%
eq_clean_data() \%>\%
eq_location_clean() \%>\%
dplyr::filter(COUNTRY \%in\% c("USA", "CHINA", "AFGHANISTAN")) \%>\%
dplyr::filter(DATE > "2000-01-01") \%>\%
ggplot() +
geom_timeline(aes(x = DATE,
                  y = COUNTRY,
                  color = TOTAL_DEATHS,
                  size = EQ_PRIMARY,
                  scaling_factor = 1
)
)
}

}
