% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_timeline_label.R
\name{geom_timeline_label}
\alias{geom_timeline_label}
\title{Earthquake Timeline}
\usage{
geom_timeline_label(mapping = NULL, data = NULL, stat = "identity",
  position = "identity", na.rm = FALSE, show.legend = NA,
  inherit.aes = TRUE, ...)
}
\arguments{
\item{mapping}{A set of aesthetics created by \code{aes()}.}

\item{data}{The NOAA earthquake data to be displayed.}

\item{stat}{The statistical transformation for that layer.}

\item{position}{Position adjustment.}

\item{na.rm}{Option to remove missing values.}

\item{show.legend}{Option to display legends.}

\item{inherit.aes}{Option to override the default aesthetics.}

\item{...}{Other arguments passed to \code{layer()}.}
}
\description{
\code{geom_timeline_label} should be used with \code{geom_timeline}. This geom adds a label (country, by default)
 to the top \code{n_max} earthquakes (sorted by magnitude, by default) per country in an earthquakes timeline plot.
}
\section{Aesthetics}{

\code{geom_timeline_label} can be used with the following aesthetics:
\itemize{
 \item x
 \item y
 \item label
 \item magnitude
 \item n_max
}
}

